feat: complete TODO in career service

+++ backend/src/main/java/ar/net/ut/backend/career/CareerService.java
@@ -6,11 +6,15 @@ import ar.net.ut.backend.career.dto.CareerUpdateDTO;
 import ar.net.ut.backend.career.event.CareerCreateEvent;
 import ar.net.ut.backend.career.event.CareerDeleteEvent;
 import ar.net.ut.backend.career.event.CareerUpdateEvent;
+import ar.net.ut.backend.course.repository.CourseRepository;
 import ar.net.ut.backend.enums.ResourceType;
+import ar.net.ut.backend.exception.BackendException;
 import ar.net.ut.backend.exception.impl.ResourceAlreadyExistsException;
 import ar.net.ut.backend.exception.impl.ResourceNotFoundException;
+import ar.net.ut.backend.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.ApplicationEventPublisher;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,6 +25,8 @@ import java.util.List;
 public class CareerService {
 
     private final CareerRepository careerRepository;
+    private final CourseRepository courseRepository;
+    private final UserRepository userRepository;
 
     private final CareerMapper careerMapper;
 
@@ -68,11 +74,16 @@ public class CareerService {
     @Transactional
     public void deleteCareer(Long id) {
         Career career = getById(id);
-
-        // TODO check that career doesn't have any courses
-        // TODO set career field to null on users over it
+        if (courseRepository.existsByCareerId(id)) {
+            throw new BackendException(
+                    "You must remove all courses within that career in order to delete it",
+                    HttpStatus.CONFLICT,
+                    "CAREER_DELETE_COURSE_CONFLICT"
+            );
+        }
 
         careerRepository.delete(career);
+        userRepository.unlinkUsersFromCareer(id);
 
         eventPublisher.publishEvent(new CareerDeleteEvent(career));
     }
